% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils.R
\name{channelNames}
\alias{channelNames}
\alias{channelNames<-}
\alias{channelNames,Image-method}
\alias{channelNames<-,Image-method}
\alias{channelNames,ImageList-method}
\alias{channelNames<-,ImageList-method}
\title{Getting and setting the channel names}
\usage{
channelNames(x)

channelNames(x) <- value

\S4method{channelNames}{Image}(x)

\S4method{channelNames}{Image}(x) <- value

\S4method{channelNames}{ImageList}(x)

\S4method{channelNames}{ImageList}(x) <- value
}
\arguments{
\item{x}{TODO}

\item{value}{TODO}
}
\description{
TODO
}
\details{
# TODO
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
