% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomapperShiny.R
\name{cytomapperShiny}
\alias{cytomapperShiny}
\title{Shiny app to gate cells on images}
\usage{
cytomapperShiny(object, mask, image = NULL, cell_id = NULL, img_id = NULL, ...)
}
\arguments{
\item{object}{an \code{\linkS4class{SingleCellExperiment}} object.}

\item{mask}{an \code{\linkS4class{CytoImageList}} object containing
single-channel \code{\linkS4class{Image}} objects}

\item{image}{(optional) an \code{\linkS4class{CytoImageList}} object containing single or
multi-channel \code{\linkS4class{Image}} objects (see Details)}

\item{cell_id}{character specifying the \code{colData(object)}, in which the
integer cell IDs are stored}

\item{img_id}{character specifying the \code{colData(object)}, in which the
integer image IDs are stored}

\item{...}{Parameters passed onto \code{\link{plotCells}} or \code{\link{plotPixels}}}
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
}
