% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-utils.R, R/utils.R
\name{ImageList-naming}
\alias{ImageList-naming}
\alias{channelNames}
\alias{names}
\alias{channelNames,Image-method}
\alias{channelNames<-,Image-method}
\alias{channelNames,ImageList-method}
\alias{channelNames<-,ImageList-method}
\alias{names<-,ImageList-method}
\title{Getting and setting the channel and image names}
\usage{
\S4method{channelNames}{Image}(x)

\S4method{channelNames}{Image}(x) <- value

\S4method{channelNames}{ImageList}(x)

\S4method{channelNames}{ImageList}(x) <- value

\S4method{names}{ImageList}(x) <- value

\S4method{channelNames}{Image}(x)

\S4method{channelNames}{Image}(x) <- value

\S4method{channelNames}{ImageList}(x)

\S4method{channelNames}{ImageList}(x) <- value
}
\arguments{
\item{x}{TODO}

\item{value}{TODO}
}
\description{
TODO

TODO
}
\details{
# TODO

# TODO
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}

Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
