% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-utils.R
\docType{methods}
\name{ImageList-subsetting}
\alias{ImageList-subsetting}
\alias{getImages}
\alias{setImages}
\alias{getChannels}
\alias{setChannels}
\alias{mergeChannels}
\alias{getImages,ImageList-method}
\alias{setImages<-,ImageList-method}
\alias{getChannels,ImageList-method}
\alias{setChannels<-,ImageList-method}
\title{General subsetting methods for ImageList objects}
\arguments{
\item{x, y}{ImageList objects}

\item{i}{integer, logical, character or vector of such indicating which
element(s) to replace or extract}

\item{value}{an \code{\linkS4class{ImageList}} or \code{\linkS4class{Image}}
object}
}
\description{
These getter and setter functions are used to extract, replace and merge
entries in an \code{\linkS4class{ImageList}} object.
}
\section{Setting and getting images}{

In the following code, \code{x} is an \linkS4class{ImageList} object.
The parameter \code{i} indicates the element(s) of \code{x} that should
be returned or replaced. Replacement is done by \code{value}, which takes
an ImageList or Image object. If \code{length(i) > 0}, \code{value}
has to be an ImageList object of \code{length(i)}, otherwise \code{value}
allows an ImageList object of length 1 or an Image object.

\describe{
\item{\code{getImages(x, i)}:}{Returns image(s) indicated by \code{i} of
the ImageList object \code{x}}
\item{\code{setImages(x, i) <- value}:}{Replaces the image(s) indicated by
\code{i} of the ImageList object \code{x} with \code{value}. For this,
\code{value} needs to have the same length as \code{i}}
}

These setter and getter functions are the recommended way of extracting and replacing
images in an ImageList object. Alternatively, the standard operations via `[`, `[[`,
`[<-` and `[[<-` can be performed (see \code{?"\link{List}"} for S4Vectors subsetting
functionality). However, these operations do not change element names during replacment
calls. The \code{setImages()} function makes sure that element names are replaced if
\code{value} is named or of \code{i} is a character of vector of characters.
}

\section{Getting and setting channels}{

TODO
Discuss adding channels to list
}

\section{Merging images}{

TODO
}

\section{Merging channels}{

TODO
}

\examples{
TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
}
