% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShiny.R
\name{launchShiny}
\alias{launchShiny}
\title{Shiny app to gate cells on images}
\usage{
launchShiny(object, mask, image = NULL, cell_id = NULL, img_id = NULL)
}
\arguments{
\item{object}{(optional) an \code{\linkS4class{SingleCellExperiment}} object.}

\item{mask}{(optional) an \code{\linkS4class{ImageList}} object containing
single-channel \code{\linkS4class{Image}} objects}

\item{image}{an \code{\linkS4class{ImageList}} object containing single or
multi-channel \code{\linkS4class{Image}} objects (see Details)}

\item{cell_id}{(optional) character specifying the \code{colData(object)}, in which the
integer cell IDs are stored}

\item{img_id}{(optional)}
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
}
