% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeChannels}
\alias{mergeChannels}
\title{Merging channels of two ImageList objects}
\usage{
mergeChannels(x, y)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}
}
\description{
TODO
}
\details{
# TODO
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
