% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasImages}
\alias{pancreasImages}
\title{Example ImageList object of image files}
\format{An ImageList object containing ...
\describe{
  \item{...}{...}
  \item{...}{...}
  ...
}}
\usage{
pancreasImages
}
\description{
TODO
}
\keyword{datasets}
