% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasMasks}
\alias{pancreasMasks}
\title{Example ImageList object of segmentation masks}
\format{An ImageList object containing ...
\describe{
  \item{...}{...}
  \item{...}{...}
  ...
}}
\usage{
pancreasMasks
}
\description{
TODO
}
\keyword{datasets}
