% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasSCE}
\alias{pancreasSCE}
\title{Example SingleCellExperiment object}
\format{A SingleCellExperiment object containing ...
\describe{
  \item{colData}{cell-level metadata...}
  \item{rowData}{marker-level metadata}
  ...
}}
\usage{
pancreasSCE
}
\description{
TODO
}
\keyword{datasets}
