% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{Plotting function to visualize cell-level information}
\usage{
plotCells(
  data,
  mask,
  colour_by = NULL,
  outline_by = NULL,
  save_image = FALSE,
  return_image = FALSE,
  col = NULL
)
}
\arguments{
\item{data}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{mask}{TODO}

\item{colour_by}{Either rownames (markers) or colnames of the colData(data) data.frame}

\item{outline_by}{TODO}

\item{save_image}{TODO}

\item{return_image}{TODO}

\item{col}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
# col = list(cell_type = c("CD4" = "red", "CD8" = "blue"),

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
