% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPixels.R
\name{plotPixels}
\alias{plotPixels}
\title{Plotting function to visualize pixel-level information}
\usage{
plotPixels(
  data,
  image,
  mask = NULL,
  colour_by,
  outline_by = NULL,
  save_image = FALSE,
  return_image = FALSE,
  col = NULL
)
}
\arguments{
\item{data}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{image}{TODO}

\item{mask}{TODO}

\item{colour_by}{TODO}

\item{outline_by}{TODO}

\item{save_image}{TODO}

\item{return_image}{TODO}

\item{col}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
