% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-param.R
\name{plotting-param}
\alias{plotting-param}
\title{Further plotting parameters for the SingleCellMapper package}
\arguments{
\item{missing_colour}{TODO}

\item{scale_bar}{TODO}

\item{image_title}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\section{Setting further parameters}{

TODO
# legend list(title, size) also allow NULL
# return_plot TODO
# return_images TODO
# save_images TODO
}

\examples{
# TODO

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})

Nicolas Damond (\email{nicolas.damond@dqbm.uzh.ch})
}
