% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrReaders.R
\name{readZARR}
\alias{readZARR}
\title{Read data from ZARR files}
\usage{
readZARR(
  file,
  type = c("omengff", "spatialdata"),
  what = c("images", "masks"),
  resolution = NULL,
  x = NULL,
  y = NULL,
  c = NULL,
  z = 1,
  t = 1,
  fov_names = NULL
)
}
\arguments{
\item{file}{TODO}

\item{type}{TODO}

\item{what}{TODO}

\item{resolution}{TODO}

\item{x}{TODO}

\item{y}{TODO}

\item{c}{TODO}

\item{z}{TODO}

\item{t}{TODO}

\item{fov_name}{TODO}
}
\value{
A \linkS4class{CytoImageList} object
}
\description{
Reads in data from a ZARR file
}
\examples{
# TODO

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch}),
}
