% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageList-utils.R
\docType{methods}
\name{ImageList-manipulation}
\alias{ImageList-manipulation}
\alias{scaleImages}
\alias{scaleImages,ImageList-method}
\alias{normalizeImages}
\alias{normalizeImages,ImageList-method}
\title{Manipulating ImageList objects}
\description{
Methods to change pixel values in ImageList objects
}
\section{Image scaling}{

In the following code, \code{x} is an \linkS4class{ImageList} object
containing one or multiple channels. In some cases, images need to be scaled
by a constant (e.g. 2^16-1 = 65535) \code{value} to revert them back to the
original pixel values after reading them in.
\describe{
\item{\code{scaleImages(x, value)}:}{Scales all images in the
\linkS4class{ImageList} object \code{x} by \code{value}
}}
}

\section{Image normalization}{

TODO
}

\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
}
