% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageList-utils.R
\docType{methods}
\name{ImageList-manipulation}
\alias{ImageList-manipulation}
\alias{scaleImages}
\alias{scaleImages,ImageList-method}
\alias{normalizeImages}
\alias{normalizeImages,ImageList-method}
\title{Manipulating ImageList objects}
\description{
Methods to change pixel values in ImageList objects
}
\section{NA}{

TODO
}

\section{Image normalization}{

TODO
}

\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
}
