% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-utils.R
\docType{methods}
\name{ImageList-naming}
\alias{ImageList-naming}
\alias{channelNames,ImageList-method}
\alias{channelNames<-,ImageList-method}
\alias{names,ImageList-method}
\alias{names<-,ImageList-method}
\title{Getting and setting the channel and image names}
\description{
TODO
}
\section{Getters}{

In the following section, \code{x} is an \linkS4class{ImageList} object and
\code{value} represents the names of either individual channels or images.
\describe{
\item{\code{channelNames(x)}:}{Returns the names of all channels stored in
\code{x}}
\item{\code{channelNames(x) <- value}:}{Replaces the channel names of
\code{x} with \code{values}. For this, \code{value} needs to have the same
length as the number of channels in \code{x}}
\item{\code{names(x)}:}{Returns the names of all images stored in
\code{x}}
\item{\code{names(x) <- value}:}{Replaces the image names of
\code{x} with \code{values}. For this, \code{value} needs to have the same
length as \code{x}}
}
}

\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
