% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subsetting-utils.R
\name{getImages}
\alias{getImages}
\alias{setImages<-}
\alias{getChannels}
\alias{setChannels<-}
\alias{ImageList-subsetting}
\alias{getImages,ImageList-method}
\alias{setImages<-,ImageList-method}
\alias{getChannels,ImageList-method}
\alias{setChannels<-,ImageList-method}
\alias{[<-,ImageList,ANY,ANY,ANY-method}
\alias{[[<-,ImageList,ANY,ANY-method}
\title{Getting and setting individual channels and images}
\usage{
getImages(x, i)

setImages(x, i) <- value

getChannels(x, i)

setChannels(x, i) <- value

\S4method{getImages}{ImageList}(x, i)

\S4method{setImages}{ImageList}(x, i) <- value

\S4method{getChannels}{ImageList}(x, i)

\S4method{setChannels}{ImageList}(x, i) <- value

\S4method{[}{ImageList,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[[}{ImageList,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{TODO}

\item{i, j}{TODO}

\item{value}{TODO}

\item{..}{TODO}
}
\value{
An ImageList object
}
\description{
These getter and setter functions are used to extracting and replacing
entries in an \code{\linkS4class{ImageList}} object.
}
\details{
These methods are preferred over subsetting via `[<-` or `[[<-`, which by default
does not replace the entry names.

It is only allowed to replace explicitily named entried in named ImageList
}
\section{Getting and setting images}{

TODO
}

\section{Getting and setting channels}{

TODO
Discuss adding channels to list
}

\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
