% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageList-class.R, R/ImageList.R
\docType{class}
\name{ImageList-class}
\alias{ImageList-class}
\alias{ImageList}
\alias{coerce,ANY,ImageList-method}
\alias{coerce,list,ImageList-method}
\alias{show,ImageList-method}
\title{S4 class for list of images}
\usage{
ImageList(...)
}
\arguments{
\item{...}{Named Image objects or a list of Image objects.}

\item{x}{TODO}

\item{value}{TODO}
}
\value{
An ImageList object
}
\description{
This class facilitates the handling of multiple one- or multi-channel images. It
inherits from \code{\linkS4class{SimpleList}} setting
\code{elementType="Image"}. Therefore, each slot contains an either one- or
multi-dimensional array in form of an \code{\linkS4class{Image}} object.
}
\details{
Similar to the \code{\linkS4class{Image}} class, the first two dimensions of each
entry indicate the spatial dimension of the image. These can be different for
each entry. The third dimension indicates the number of channels per
Image. Each entry in the ImageList class object must contain the same
number of channels. Here, each channel represents pixel values indicating
measurement intensities or in case of segmentation masks the cells' ID.
The ImageList class therefore only supports a Grayscale colormode (see
\code{\link[EBImage]{colormode}}) representation of each individual image.

The class further contains an \code{\link{elementMetadata}} slot that stores
image-level meta information. This slot should be accessed using the
\code{\link[S4Vectors]{mcols}} accessor function.
}
\section{Restrictions on entries}{

Discuss named list - not empty names, no NA, not duplicated names
}

\section{Accessors}{

channelNames
mcols
}

\section{Coercion}{

An \code{ImageList} object can be coerced from \code{list()},
\code{SimpleList} and \code{List} objects. Also \code{list},
\code{SimpleList} and \code{List} objects can be coerced to \code{ImageList}.
}

\section{Subsetting}{

getImages
setImages
mergeChannels
}

\section{Looping}{

endoapply, mendoapply
}

\examples{
# Creation of ImageList

# Accessors

# Coercion
# as("ImageList", list)

# Subsetting

# Looping

}
\seealso{
\code{\linkS4class{Image}}, for ..

\code{\linkS4class{SimpleList}}, for ..
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}.
}
