% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/ImageList.R
\docType{class}
\name{ImageList-class}
\alias{ImageList-class}
\alias{ImageList}
\title{S4 class for list of images}
\usage{
ImageList(..., elementMetadata = NULL)
}
\arguments{
\item{...}{TODO}

\item{elementMetadata}{TODO}
}
\description{
This class allows the handling of multiple one- or multi-channel images.
The class inherits from the \linkS4class{SimpleList} object class storing \linkS4class{Image} objects in each slot.
Furthermore, the class contains an \code{elementMetadata} slot that stores image-level meta information.
}
\details{
In this class ...
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch}
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
