% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadImages.R
\name{loadImages}
\alias{loadImages}
\title{Function to read in images}
\usage{
loadImages(x, pattern = NULL, ...)
}
\arguments{
\item{x}{The function takes a variety of possible character inputs
\describe{
\item{A single file}{}.
\item{A path}{}.
\item{A list of files}{}.
}}

\item{pattern}{character vector of the following form
\describe{
\item{A single character specifying a pattern to search for in the specified path
(regular expressions are supported)}{}.
\item{A character vector in which unique entries are matched against
file names in the specified path}{}.
}}

\item{...}{arguments passed to the \code{\link{readImage}} function}
}
\value{
An \linkS4class{Image} or \linkS4class{ImageList} object
}
\description{
Function to read in single- or multi-channel images
}
\examples{
# Providing a single file
single.image <- system.file("extdata/A02_mask.tiff", package = "SingleCellMapper")
single.image <- loadImages(single.image)

# Providing a path and pattern
path.to.images <- system.file("extdata", package = "SingleCellMapper")
image.list <- loadImages(path.to.images, pattern = "mask.tiff")

# Providing multiple patterns
data(pancreasSCE)
path.to.images <- system.file("extdata", package = "SingleCellMapper")
image.list <- loadImages(path.to.images, pattern = pancreasSCE$MaskName)

}
\seealso{
\code{\link{readImage}}, for reading in individual images.
}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
