% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasImages}
\alias{pancreasImages}
\title{Example CytoImageList object of image files}
\format{
A CytoImageList object containing 3 \code{\linkS4class{Image}}
objects with 5 channels each. Channel names can be accessed via
\code{?\link{channelNames}}.
}
\usage{
pancreasImages
}
\description{
This \linkS4class{CytoImageList} object contains multi-channel stacks of
three images acquired by imaging mass cytometry. Each channel represents the
pixel-intensities of each of the 5 measured proteins. The data is part of a
imaging mass cytometry study on the progression of Type 1 diabetes and
contains pancreas cells.
}
\references{
\href{https://www.sciencedirect.com/science/article/pii/S1550413118306910}{
Damond, N. et al., A Map of Human Type 1 Diabetes Progression by
Imaging Mass Cytometry, Cell Metabolism 29:3, 2019}
}
\keyword{datasets}
