% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{Plotting function to visualize cell-level information}
\usage{
plotCells(
  object,
  mask,
  cell_ID,
  image_ID,
  colour_by = NULL,
  outline_by = NULL,
  exprs_values = "counts",
  subset_images = NULL,
  save_images = NULL,
  return_images = FALSE,
  colour = NULL,
  missing_colour = "gray",
  scale_bar = list(length = 20, label = NULL, lwd = 2, col = "white", margin = 10),
  ...
)
}
\arguments{
\item{object}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{mask}{TODO}

\item{cell_ID}{TODO}

\item{image_ID}{TODO}

\item{colour_by}{TODO}

\item{outline_by}{TODO}

\item{exprs_values}{TODO}

\item{subset_images}{TODO}

\item{save_images}{TODO}

\item{return_images}{TODO}

\item{colour}{TODO}

\item{missing_colour}{TODO}

\item{scale_bar}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\section{Linking image IDs and cell IDs}{

TODO
}

\examples{
# TODO
# col = list(cell_type = c("CD4" = "red", "CD8" = "blue"),

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})

Nicolas Damond (\email{nicolas.damond@dqbm.uzh.ch})
}
