% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{}
\usage{
plotCells(
  data,
  mask,
  colour_by = cell_type,
  outline_by = tumour_stroma,
  save_image = FALSE,
  return_image = FALSE,
  col = list(cell_type = c(CD4 = "red", CD8 = "blue"), tumour_stroma = c(tumour =
    "white", stroma = "black"))
)
}
\arguments{
\item{data}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{mask}{TODO}

\item{colour_by}{Either rownames (markers) or colnames of the colData(data) data.frame}

\item{outline_by}{TODO}

\item{save_image}{TODO}

\item{return_image}{TODO}

\item{col}{TODO}
}
\value{

}
\description{

}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
