% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCells.R
\name{plotCells}
\alias{plotCells}
\title{Function to visualize cell-level information on segmentation masks}
\usage{
plotCells(
  object,
  mask,
  cell_id,
  img_id,
  colour_by = NULL,
  outline_by = NULL,
  exprs_values = "counts",
  subset_images = NULL,
  colour = NULL,
  missing_colour = "gray",
  scale_bar = list(length = 20, label = NULL, cex = 1, lwd = 2, colour = "white", margin = c(10,10)),
  image_title = list(text =  NULL, position = "top", cex = 1, colour = "white", margin = 10, font = 2),
  ...
)
}
\arguments{
\item{object}{a \code{\linkS4class{SingleCellExperiment}} object.}

\item{mask}{an \code{\linkS4class{ImageList}} object single-channel
\code{\linkS4class{Image}} objects (see Details)}

\item{cell_id}{character specifying the \code{colData(object)}, in which the
integer cell IDs are stored}

\item{img_id}{TODO}

\item{colour_by}{TODO}

\item{outline_by}{TODO}

\item{exprs_values}{TODO}

\item{subset_images}{TODO}

\item{colour}{TODO}

\item{missing_colour}{TODO}

\item{scale_bar}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
This function takes a \code{\linkS4class{SingleCellExperiment}} and
\code{\linkS4class{ImageList}} object to colour cells by marker expression or
metadata.
}
\section{Segmentaion mask object}{

TODO
}

\section{Linking image IDs and cell IDs}{

TODO
}

\examples{
# TODO
# colour = list(cell_type = c("CD4" = "red", "CD8" = "blue"),

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})

Nicolas Damond (\email{nicolas.damond@dqbm.uzh.ch})
}
