% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPixels.R
\name{plotPixels}
\alias{plotPixels}
\title{Function to visualize pixel-level information of multi-channel images}
\usage{
plotPixels(
  image,
  object = NULL,
  mask = NULL,
  cell_id = NULL,
  img_id = NULL,
  colour_by = NULL,
  bcg = NULL,
  outline_by = NULL,
  subset_images = NULL,
  colour = NULL,
  ...
)
}
\arguments{
\item{image}{an \code{\linkS4class{ImageList}} object containing single or
multi-channel \code{\linkS4class{Image}} objects (see Details)}

\item{object}{(optional) an \code{\linkS4class{SingleCellExperiment}} object.}

\item{mask}{(optional) an \code{\linkS4class{ImageList}} object containing
single-channel \code{\linkS4class{Image}} objects}

\item{cell_id}{(optional) character specifying the \code{colData(object)}, in which the
integer cell IDs are stored}

\item{img_id}{(optional)}

\item{colour_by}{TODO}

\item{bcg}{brightness/contrast/gamma list("H3" = c(0,1,1))}

\item{outline_by}{(optional)}

\item{subset_images}{TODO Note that Masks and Images need to have same names or same IDs}

\item{colour}{TODO}

\item{...}{Further arguments passed to  \code{?"\link{plotting-param}"}}
}
\value{
TODO
}
\description{
This function takes a \code{\linkS4class{ImageList}} object to colour pixels
by marker expression. Additionally, a \code{\linkS4class{SingleCellExperiment}}
object and \code{\linkS4class{ImageList}} object containing segmentation masks
can be provided to outline cells based on metadata.
}
\examples{
# TODO

}
\author{
Nils Eling \email{nils.eling@dqbm.uzh.ch},
Nicolas Damond \email{nicolas.damond@dqbm.uzh.ch}
}
