% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-param.R
\name{plotting-param}
\alias{plotting-param}
\title{Further plotting parameters for the SingleCellMapper package}
\arguments{
\item{missing_colour}{TODO}

\item{background_colour}{TODO}

\item{scale_bar}{TODO}

\item{image_title}{TODO}

\item{save_image}{TODO}

\item{return_plot}{TODO}

\item{return_images}{TODO}

\item{legend}{TODO}

\item{margin}{TODO}

\item{display}{TODO}

\item{scale}{TODO}

\item{interpolate}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\section{Setting further parameters}{

TODO
}

\examples{
# TODO

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})

Nicolas Damond (\email{nicolas.damond@dqbm.uzh.ch})
}
