% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-param.R
\name{plotting-param}
\alias{plotting-param}
\title{Further plotting parameters for the SingleCellMapper package}
\arguments{
\item{legend_size}{TODO}

\item{title_size}{TODO}

\item{title_colour}{TODO}

\item{return_plot}{TODO}

\item{return_images}{TODO}

\item{save_images}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\section{Setting further parameters}{

TODO
}

\examples{
# TODO

}
\author{
Nils Eling (\email{nils.eling@dqbm.uzh.ch})

Nicolas Damond (\email{nicolas.damond@dqbm.uzh.ch})
}
